ContinualTrainer implemented
