+++ continualnat/models/cmlm/cmlm.py
@@ -142,7 +142,7 @@ class CMLM(TransformerNATCore):
         references = batch["references"]
 
         # Compute translations
-        tokenizer, lang_pair, tgt_lang = self._val_tokenizer_tgtlang(dataloader_idx)
+        tokenizer, lang_pair, tgt_lang = super()._val_tokenizer_tgtlang(dataloader_idx)
         translation, _ = self.generate(input_ids, tokenizer.lang_code_to_id[tgt_lang])
         predictions = tokenizer.batch_decode(translation, skip_special_tokens=True)
